idek anymore

+++ server/controllers/tripsController.js
@@ -1,23 +1,34 @@
 import { pool } from '../config/database.js'
 
 const createTrip = async (req, res) => {
+    if (!req.user) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
+    
     try {
         const { title, description, img_url, num_days, start_date, end_date, total_cost } = req.body
         const results = await pool.query (
-            `INSERT INTO trips (id, title, description, img_url, num_days, start_date, end_date, total_cost)
-            VALUES (DEFAULT, $1, $2, $3, $4, $5, $6, $7)
+            `INSERT INTO trips (user_id, title, description, img_url, num_days, start_date, end_date, total_cost)
+            VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
             RETURNING *`,
-            [title, description, img_url, num_days, start_date, end_date, total_cost]
+            [req.user.id, title, description, img_url, num_days, start_date, end_date, total_cost]
         )
         res.status(201).json(results.rows[0])
     } catch (error) {
+        console.error("Error creating trip:", error.message)
         res.status(409).json({ message: error.message })
     }
 }
 
-const getTrips = async (req, res) => {
+const getTripsbyUser = async (req, res) => {
+    if (!req.user) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
+
     try {
-        const results = await pool.query(`SELECT * FROM trips ORDER BY id ASC`)
+        const results = await pool.query(`SELECT * FROM trips WHERE user_id = $1`,
+            [req.user.id]
+        )
         res.status(200).json(results.rows)
     } catch(error) {
         res.status(409).json( { error: error.message } )
@@ -25,9 +36,15 @@ const getTrips = async (req, res) => {
 }
 
 const getTrip = async (req, res) => {
+    if (!req.user) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
+
+    const userId = req.user.id
+    const tripId = req.params.id
+
     try {
-        const id = parseInt(req.params.id);
-        const results = await pool.query(`SELECT * FROM trips WHERE id = $1`, [id])
+        const results = await pool.query(`SELECT * FROM trips WHERE id = $1 AND user_id = $2`, [tripId, userId])
         res.status(200).json(results.rows[0])
     } catch(error) {
         res.status(409).json( { error: error.message } )
@@ -37,15 +54,21 @@ const getTrip = async (req, res) => {
 }
 
 const updateTrip = async (req, res) => {
+    if (!req.user) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
+
+    const userId = req.user.id
+    const tripId = req.params.id
+
     try {
         const { title, description, img_url, num_days, start_date, end_date, total_cost } = req.body
-        const id = parseInt(req.params.id)
     
         const results = await pool.query(
           `UPDATE trips
           SET title = $1, description = $2, img_url = $3, num_days = $4, start_date = $5, end_date = $6, total_cost= $7
-          WHERE id = $8`,
-          [title, description, img_url, num_days, start_date, end_date, total_cost, id]
+          WHERE id = $8 AND user_id = $9`,
+          [title, description, img_url, num_days, start_date, end_date, total_cost, tripId, userId]
         )
     
         res.status(200).json(results.rows);
@@ -55,12 +78,18 @@ const updateTrip = async (req, res) => {
     }
 
 const deleteTrip = async (req, res) => {
-    const id = parseInt(req.params.id)
+    if (!req.user) {
+        return res.status(401).json({ message: 'Unauthorized' })
+    }
+    
+    const userId = req.params.id
+    const tripId = req.user.id
+
     try {
       const activity_deletion = await pool.query(
-        `DELETE FROM activities
-        WHERE trip_id = $1`,
-        [id]
+        `DELETE FROM trips
+        WHERE id = $1 AND user_id = $2 RETURNING *`,
+        [tripId, userId]
       )
       const results = await pool.query('DELETE FROM trips WHERE id = $1', [id])
       res.status(200).json(results.rows)
@@ -80,7 +109,7 @@ const deleteAllTrips = async (req, res) => {
 
 export default {
     createTrip,
-    getTrips,
+    getTripsbyUser,
     getTrip,
     updateTrip,
     deleteTrip,
